Add HW 04-auth

+++ controllers/contactsControllers.js
@@ -3,14 +3,16 @@ import HttpError from "../helpers/HttpError.js";
 import controllerWrapper from "../helpers/controllerWrapper.js";
 
 const listContacts = async (req, res) => {
-    const result = await contactsService.listContacts();
+    const {id} = req.user;
+    const result = await contactsService.listContacts({owner: id});
 
     res.json(result);
 };
 
 const getContactById = async (req, res) => {
     const {contactId} = req.params;
-    const result = await contactsService.getContactById(contactId);
+    const {id} = req.user;
+    const result = await contactsService.getContactById({contactId: contactId, owner: id});
 
     if (!result) {
         throw HttpError(404);
@@ -21,7 +23,8 @@ const getContactById = async (req, res) => {
 
 const removeContact = async (req, res) => {
     const {contactId} = req.params;
-    const result = await contactsService.removeContact(contactId);
+    const {id} = req.user;
+    const result = await contactsService.removeContact({contactId: contactId, owner: id});
 
     if (!result) {
         throw HttpError(404);
@@ -31,7 +34,7 @@ const removeContact = async (req, res) => {
 };
 
 const createContact = async (req, res) => {
-    const result = await contactsService.addContact(req.body);
+    const result = await contactsService.addContact({...req.body, owner: req.user.id});
 
     res.status(201).json(result);
 };
@@ -41,8 +44,9 @@ const updateContact = async (req, res) => {
         throw HttpError(400, "Body must have at least one field");
     }
 
-    const {contactId} = req.params;
-    const result = await contactsService.updateContact(contactId, req.body);
+    const {id} = req.params;
+    const {owner} = req.user;
+    const result = await contactsService.updateContact({id, owner}, req.body);
 
     if (!result) {
         throw HttpError(404);
